pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.11.0,<3.12"
flask = "^3.0.0"
gunicorn = "^21.2.0"
flask-sqlalchemy = "^3.1.1"
flask-wtf = "^1.2.2"
wtforms = "^3.2.1"
flask-bcrypt = "^1.0.1"
flask-login = "^0.6.3"
email-validator = "2.2.0"
wtforms-sqlalchemy = "^0.4.2"
flask-migrate = "4.0.7"
sqlalchemy = "^2.0.41"
cloudinary = "^1.44.1"
gspread = "6.1.4"
google-auth-oauthlib = "1.2.1"
google-auth = "^2.40.3"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"